orbit_propagation/requirements.txt
astropy==6.0.0
numpy==1.26.4
pandas==2.2.1
poliastro==0.17.0
python_dateutil==2.8.2
scipy==1.12.0
sgp4==2.23
spacetrack==1.2.0
